remove createModelAlcoState to handlers
